Project 3(COMP3334_Project-main)/requirements.txt
certifi==2021.10.8
dnspython==2.2.1
Pillow==9.1.0
pymongo==4.1.1
tzdata==2022.1
